Companion for #substrate-8955

+++ runtime/kusama/src/lib.rs
@@ -156,7 +156,7 @@ type MoreThanHalfCouncil = EnsureOneOf<
 
 parameter_types! {
 	pub const Version: RuntimeVersion = VERSION;
-	pub const SS58Prefix: u8 = 2;
+	pub const SS58Prefix: u16 = 2;
 }
 
 impl frame_system::Config for Runtime {
