+++ runtime/rococo/src/lib.rs
@@ -269,7 +269,7 @@ impl Contains<Call> for BaseFilter {
 
 parameter_types! {
 	pub const Version: RuntimeVersion = VERSION;
-	pub const SS58Prefix: u8 = 42;
+	pub const SS58Prefix: u16 = 42;
 }
 
 impl frame_system::Config for Runtime {
