+++ runtime/test-runtime/src/lib.rs
@@ -122,7 +122,7 @@ sp_api::decl_runtime_apis! {
 
 parameter_types! {
 	pub const Version: RuntimeVersion = VERSION;
-	pub const SS58Prefix: u8 = 42;
+	pub const SS58Prefix: u16 = 42;
 }
 
 impl frame_system::Config for Runtime {
